extreme values
Former-commit-id: 97a1585246fc1dbad15c9ccbfd6a7f766e357d38

+++ packages/lambda/src/io.ts
@@ -23,9 +23,12 @@ const ensureDir = async ({bucketName}: {bucketName: string}) => {
 	}
 };
 
-export const lambdaLs = async (bucketName: string): Promise<_Object[]> => {
+export const lambdaLs = async (
+	bucketName: string,
+	forceS3: boolean
+): Promise<_Object[]> => {
 	await ensureDir({bucketName});
-	if (ENABLE_EFS) {
+	if (ENABLE_EFS && !forceS3) {
 		const dir = await promises.readdir(EFS_MOUNT_PATH + '/' + bucketName);
 
 		// TODO: Mock data
@@ -53,14 +56,16 @@ export const lambdaWriteFile = async ({
 	bucketName,
 	key,
 	body,
+	forceS3,
 }: {
 	bucketName: string;
 	key: string;
 	body: ReadStream | string;
+	forceS3: boolean;
 }): Promise<void> => {
 	await ensureDir({bucketName});
 
-	if (ENABLE_EFS) {
+	if (ENABLE_EFS && !forceS3) {
 		if (typeof body === 'string') {
 			return promises.writeFile(
 				EFS_MOUNT_PATH + '/' + bucketName + '/' + key,
